Use central control logic for hui-card and more-info
